Dev replace below confidence (#32)
* replace_below_confidence when lower confidence

The intent is to reset data to some default value when
the confidence level is below certain value.

Sample usage:

```
    replace_below_confidence($interval, 1.0, 0.0)
'''

In the above case 1.0 is the threshold confidence level,
below which the metric value will be reset to 0.0 (3rd
argument).

@see https://github.com/dalmatinerdb/dalmatinerdb/issues/134

* threshold equal or lower values to default

As per review comment allow query like `reset(<query>, 0, 0)` to
reset datapoints with 0 confidence.

@see https://github.com/dalmatinerdb/dalmatinerdb/issues/134

* basic test for replace_below_confidence

There is no public API to directly change confidence value,
so this is ideally not a true test. Having said that it
could have been done with a mix of using other methods.
Since, those methods are in another moule like mmath_aggr,
so I dont want to create its dependency here.